Enable different ways of rendering frames: * Traditionally, with a single bitmapData for every frame * By drawing a rect from one single bitmap data * Using tilesheets
For this render targets are introduced, which render into a "graphics" object.
For now the graphics object is the "graphics" member of the AnimatedSprite.

Also: Introduced flag for rendering with "ADD" blendMode.
Due to some issue, rendering with blendmode "Add" works only using tilesheets on C++
targets and by setting "blendMode" of the sprite on flash target.