Discard sourceImage when all frames generated.
After the final frame is pulled in by generateBitmap(i) call, it appears
we can safely null both sourceImage and sourceImageAlpha.

The usual uses of merge and updateImage and other methods seems to be
supported; the one case this might break is a "merge with another
spritesheet with the same sourceImage and sourceImageAlpha as ourselves",
causing it to generate frames earlier than it would have otherwise.